AUT-4588: Move the CONTACT_FORM_STRUCTURE into a dedicated structure contact-us-structure file
AUT-4588: Add new type to contact-us-structure
